// src/types/contentful.d.ts
export interface IAccordion extends Entry<IAccordionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "accordion";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICareersList extends Entry<ICareersListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "careersList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IComparisonTableFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Theme */
  theme?: "dark" | "light" | undefined;

  /** Headline */
  headline?: IHeadline | undefined;

  /** Section ID */
  sectionId?: string | undefined;

  /** Table */
  table?: Document | undefined;
}

export interface IComparisonTable extends Entry<IComparisonTableFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "comparisonTable";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICtaBanner extends Entry<ICtaBannerFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ctaBanner";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IDividerFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Theme */
  theme?: "dark" | "light" | undefined;
}

/** A horizontal line to divide up content and sections */

export interface IDivider extends Entry<IDividerFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "divider";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IFeature extends Entry<IFeatureFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "feature";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IFeatureStack extends Entry<IFeatureStackFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "featureStack";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IHeadline extends Entry<IHeadlineFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "headline";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IHero extends Entry<IHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "hero";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IInPageNav extends Entry<IInPageNavFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "inPageNav";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ILinkFields {
  /** Internal Name */
  name?: string | undefined;

  /** Label */
  text: string;

  /** Page */
  page?: IPage | IPageSimple | IPost | undefined;

  /** URL */
  url?: string | undefined;

  /** New Window */
  newWindow?: boolean | undefined;
}

/** Call to Action */

export interface ILink extends Entry<ILinkFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "link";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IList extends Entry<IListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "list";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ILogoCloud extends Entry<ILogoCloudFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "logoCloud";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IMulticard extends Entry<IMulticardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "multicard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface INavItem extends Entry<INavItemFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "navItem";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageFields {
  /** Theme */
  theme?: "dark" | "light" | undefined;

  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Sections */
  sections?:
    | (
        | ISection
        | IAccordion
        | ICareersList
        | IFeature
        | IHero
        | IList
        | ILogoCloud
        | IMulticard
        | IPanorama
        | IPostList
        | IQuote
        | IText
        | IHeadline
        | ICtaBanner
        | IFeatureStack
        | IDivider
        | IVideo
        | IComparisonTable
        | IInPageNav
      )[]
    | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Display this translation? */
  translated?: boolean | undefined;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageSimpleFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Text */
  text?: Document | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Display this translation? */
  translated?: boolean | undefined;
}

export interface IPageSimple extends Entry<IPageSimpleFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pageSimple";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPanoramaFields {
  /** Internal Name */
  name?: string | undefined;

  /** Theme */
  theme?: "dark" | "light" | undefined;

  /** Headline */
  headline?: IHeadline | undefined;

  /** Section ID */
  sectionId?: string | undefined;

  /** Images */
  images: Asset[];

  /** Ratio */
  ratio?: "auto" | "1-1" | "4-5" | "16-9" | undefined;
}

export interface IPanorama extends Entry<IPanoramaFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "panorama";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPerson extends Entry<IPersonFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "person";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPostFields {
  /** Theme */
  theme?: "light" | "dark" | undefined;

  /** Title */
  title: string;

  /** Type */
  type?: "blog" | undefined;

  /** Category */
  category?: "Design" | "Engineering" | "Work" | undefined;

  /** Slug */
  slug: string;

  /** Eyebrow */
  eyebrow?: string | undefined;

  /** Subhead */
  subhead?: string | undefined;

  /** Subhead Description */
  subheadDescription?: string | undefined;

  /** Text */
  text: Document;

  /** Thumbnail Text */
  thumbnailText?: string | undefined;

  /** Author */
  author?: IPerson | undefined;

  /** Published On */
  publishedOn?: string | undefined;

  /** Hero Image */
  heroImage?: Asset | undefined;

  /** Thumbnail Image */
  thumbnailImage?: Asset | undefined;

  /** ctaBanner */
  ctaBanner?: ICtaBanner[] | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Show Table of Contents */
  toc?: boolean | undefined;

  /** Display this translation? */
  translated?: boolean | undefined;
}

export interface IPost extends Entry<IPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "post";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPostList extends Entry<IPostListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "postList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IQuote extends Entry<IQuoteFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "quote";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ISection extends Entry<ISectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "section";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ISeo extends Entry<ISeoFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "seo";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISettingsFields {
  /** Site Title */
  title: string;

  /** Legal Name */
  legalName?: string | undefined;

  /** Metadata */
  metadata?: ISeo | undefined;

  /** Logo Svg */
  logoSvg?: string | undefined;

  /** Logo Image */
  logoImage?: Asset | undefined;

  /** Image Quality */
  imagesQuality?: number | undefined;

  /** Main Navigation */
  mainNavigation?: (ILink | ISimpleList | INavItem)[] | undefined;

  /** Nav CTA 1 */
  navCta1?: ILink | undefined;

  /** Nav CTA 2 */
  navCta2?: ILink | undefined;

  /** Footer Navigation */
  footerNavigation?: (ILink | ISimpleList)[] | undefined;

  /** Legal Navigation */
  legalNavigation?: ILink[] | undefined;

  /** Facebook */
  facebook?: string | undefined;

  /** Twitter */
  twitter?: string | undefined;

  /** Instagram */
  instagram?: string | undefined;

  /** LinkedIn */
  linkedIn?: string | undefined;

  /** Pinterest */
  pinterest?: string | undefined;

  /** Google Analytics ID */
  googleAnalyticsId?: string | undefined;

  /** Google Tag Manager ID */
  googleTagManagerId?: string | undefined;

  /** CookieBot Id */
  cookieBotId?: string | undefined;

  /** Facebook App ID */
  facebookAppId?: string | undefined;

  /** Greenhouse Board ID */
  greenhouseBoardToken?: string | undefined;

  /** Marker.io Destination ID */
  markerIoDestinationId?: string | undefined;

  /** Marketo Hostname */
  marketoHostname?: string | undefined;

  /** Marketo Munchkin ID */
  marketoMunchkinId?: string | undefined;

  /** Enable Search */
  enableSearch?: boolean | undefined;
}

export interface ISettings extends Entry<ISettingsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "settings";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISimpleListFields {
  /** Internal Name */
  name?: string | undefined;

  /** Title */
  title?: string | undefined;

  /** Items */
  items?: ILink[] | undefined;
}

/** Specific to navigation. Used for sub-navigation list */

export interface ISimpleList extends Entry<ISimpleListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "simpleList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IText extends Entry<ITextFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "text";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IVideo extends Entry<IVideoFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "video";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}